Fix errors, test-typeCheck move to instrument test.
