+++ laevatein/build.gradle
@@ -6,6 +6,8 @@ android {
     defaultConfig {
         minSdkVersion 14
         targetSdkVersion 26
+
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
     compileOptions {
         sourceCompatibility JavaVersion.VERSION_1_7
@@ -15,6 +17,8 @@ android {
         abortOnError false
     }
     testOptions.unitTests.includeAndroidResources true
+
+    useLibrary 'android.test.mock'
 }
 
 dependencies {
@@ -36,6 +40,8 @@ dependencies {
     testImplementation 'org.robolectric:robolectric:3.5.1'
     testImplementation 'org.robolectric:shadows-support-v4:3.3.2'
     testImplementation 'org.mockito:mockito-all:1.10.19'
+
+    androidTestImplementation 'androidx.test:runner:1.1.1'
 }
 
 apply from: 'https://raw.github.com/chrisbanes/gradle-mvn-push/master/gradle-mvn-push.gradle'
